payment-system/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis[hiredis]==5.2.1

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
